Example 19.5 One way to deal with login errors.
Through a lot of trial and error, this method still connects to gmail (reconnect_max_attempts was required for that) and deals with bad servers, bad usernames, and invalid credentials. It just prints an error to the console, but it would be trivial to update it to show an error message in a popup.

However, running this code locks up the interface in an unappealing way while it waits for a connection and requests the buddy list. While this happens, it is not possible to interact with the buttons or text boxes. So let's try a different solution.